services/ai-engine/requirements.txt
# VITAL Phase 2 Expert Service Requirements
# Cloud deployment with real Supabase and Pinecone integration

# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.9.0  # Updated to fix langsmith compatibility
pydantic-settings==2.5.0  # Required for BaseSettings in pydantic 2.x

# LangChain Ecosystem - Pinned compatible versions (TRULY FINAL!)
# Critical: langgraph 0.2.28 requires langgraph-checkpoint <2.0.0 (must be 1.x!)
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16
langchain-core==0.2.39
langgraph==0.2.28
langgraph-checkpoint==1.0.12
langgraph-checkpoint-sqlite==1.0.3
langsmith==0.1.129

# Database & Vector Store
supabase==2.3.0
pinecone==7.3.0  # Fixed: was pinecone-client==2.2.4 (old package)

# HTTP & Utilities
httpx>=0.24.0,<0.25.0  # Compatible with supabase==2.3.0
aiohttp==3.9.1  # Async HTTP for web tools
python-dotenv==1.0.0
numpy>=1.26.0  # Updated for Python 3.13 compatibility

# Web Scraping & HTML Parsing
beautifulsoup4==4.12.2
lxml>=5.3.0  # For faster HTML parsing (Python 3.13 compatible)
html5lib==1.1  # Alternative HTML parser

# Text Processing & Language Detection
unidecode==1.3.7  # Unicode normalization for filenames
langdetect==1.0.9  # Language detection (better than simple implementation)
python-dateutil==2.8.2  # Enhanced date parsing
sentence-transformers>=2.7.0  # Semantic embeddings for Phase 2 Memory (Python 3.13 compatible)
faiss-cpu>=1.9.0  # Fast similarity search (Python 3.13 compatible)

# Logging & Monitoring
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk[fastapi]==2.18.0  # Error tracking and performance monitoring

# Caching & Performance
redis>=5.0.0

# Resilience & Error Handling
tenacity>=8.0.0  # Retry logic
circuitbreaker>=1.4.0  # Circuit breaker pattern
slowapi>=0.1.9  # Rate limiting

# Security Dependencies
vecs==0.3.0
modal==0.62.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Shared Kernel (Multi-Tenant Core)

# Development
pytest==7.4.3
pytest-asyncio==0.21.1

# LangGraph GUI Dependencies
# Note: Using custom implementation in src/langgraph_gui/
# Official langgraph-gui package not required as we have custom integration
reportlab>=4.0.0  # For PDF generation in workflows
langchain-ollama>=0.1.0  # For Ollama LLM support (optional)
